[Feat] DefaultMyProfileRepository에 내 프로필 CRUD관련 유저디폴츠 활용한 캐싱로직 추가

+++ travelPlan/Source/Data/Repositories/DefaultUserInfoRepository.swift
@@ -10,6 +10,7 @@ import Combine
 final class DefaultMyProfileRepository {
   // MARK: - Dependencies
   private let service: Sessionable
+  private lazy var othersProfileRepository = DefaultOthersProfileRepository(service: service)
   private typealias MyInfoManager = UserDefaultsManager.User
   
   // MARK: - Properties
@@ -43,8 +44,13 @@ extension DefaultMyProfileRepository: MyProfileRepository {
   }
   
   func updateUserNickname(with name: String) -> Future<Bool, MainError> {
-    //TODO: - 로그인 중인 사용자의 userId를 키체인이나 유저디폴츠에서 가져와야 합니다.
-    let requestDTO = UserNicknamePatchRequestDTO(nickname: name, userId: 000)
+    guard let loggedInUserId = MyInfoManager.id else {
+      return Future { promise in
+        promise(.failure(MainError.referenceError(.other("유저디폴츠에서 로그인한 사용자의 id가져올 수 없습니다."))))
+      }
+    }
+    
+    let requestDTO = UserNicknamePatchRequestDTO(nickname: name, userId: loggedInUserId)
     let endpoint = UserInfoAPIEndpoint.updateUserNickname(with: requestDTO)
     return .init { [unowned self] promise in
       service.request(endpoint: endpoint)
@@ -57,14 +63,22 @@ extension DefaultMyProfileRepository: MyProfileRepository {
             promise(.failure(error))
           }
         } receiveValue: { responseDTO in
+          if responseDTO.result {
+            MyInfoManager.updateNickname(with: name)
+          }
           promise(.success(responseDTO.result))
         }.store(in: &subscriptions)
     }
   }
   
   func updateProfile(with profile: String) -> Future<Bool, MainError> {
-    // TODO: - UserDefaults 관리 담당 객체를 통해 로그인 한 사용자의 userID를 가져와야 합니다.
-    let userIdReqeustDTO = UserIdReqeustDTO(userId: 13)
+    guard let loggedInUserId = MyInfoManager.id else {
+      return Future { promise in
+        promise(.failure(MainError.referenceError(.other("유저디폴츠에서 로그인한 사용자의 id가져올 수 없습니다."))))
+      }
+    }
+    
+    let userIdReqeustDTO = UserIdReqeustDTO(userId: loggedInUserId)
     let reqeustDTO = UserProfileRequestDTO(profile: profile)
     let endpoint = UserInfoAPIEndpoint.updateProfile(withQuery: userIdReqeustDTO, body: reqeustDTO)
     return Future<Bool, MainError> { [unowned self] promise in
@@ -76,14 +90,20 @@ extension DefaultMyProfileRepository: MyProfileRepository {
           }
         } receiveValue: { responseDTO in
           let isSucceed = (200...299).contains(Int(responseDTO.status) ?? -1)
+          MyInfoManager.updateProfileURL(with: responseDTO.result.imageURL)
           promise(.success(isSucceed))
         }.store(in: &subscriptions)
     }
   }
   
   func saveProfile(with profile: String) -> Future<Bool, MainError> {
-    // TODO: - UserDefaults 관리 담당 객체를 통해 로그인 한 사용자의 userID를 가져와야 합니다.
-    let userIdRequestDTO = UserIdReqeustDTO(userId: 13)
+    guard let loggedInUserId = MyInfoManager.id else {
+      return Future { promise in
+        promise(.failure(MainError.referenceError(.other("유저디폴츠에서 로그인한 사용자의 id가져올 수 없습니다."))))
+      }
+    }
+    
+    let userIdRequestDTO = UserIdReqeustDTO(userId: loggedInUserId)
     let requestDTO = UserProfileRequestDTO(profile: profile)
     let endpoint = UserInfoAPIEndpoint.saveProfile(withQuery: userIdRequestDTO, body: requestDTO)
     return Future<Bool, MainError> { [unowned self] promise in
@@ -101,8 +121,13 @@ extension DefaultMyProfileRepository: MyProfileRepository {
   }
   
   func deleteProfile() -> Future<Bool, MainError> {
-    // TODO: - UserDefaults 관리 담당 객체를 통해 로그인 한 사용자의 userID를 가져와야 합니다.
-    let requestDTO = UserIdReqeustDTO(userId: 13)
+    guard let loggedInUserId = MyInfoManager.id else {
+      return Future { promise in
+        promise(.failure(MainError.referenceError(.other("유저디폴츠에서 로그인한 사용자의 id가져올 수 없습니다."))))
+      }
+    }
+    
+    let requestDTO = UserIdReqeustDTO(userId: loggedInUserId)
     let endpoint = UserInfoAPIEndpoint.deleteProfile(with: requestDTO)
     return Future<Bool, MainError> { [unowned self] promise in
       service.request(endpoint: endpoint)
@@ -112,25 +137,36 @@ extension DefaultMyProfileRepository: MyProfileRepository {
             promise(.failure(error))
           }
         } receiveValue: { responseDTO in
+          MyInfoManager.deleteProfile()
           promise(.success(responseDTO.result))
         }.store(in: &subscriptions)
     }
   }
   
   func fetchProfile() -> Future<ProfileImageEntity, MainError> {
-    // TODO: - UserDefaults 관리 담당 객체를 통해 로그인 한 사용자의 userID를 가져와야 합니다.
-    let requestDTO = UserIdReqeustDTO(userId: 13)
-    let endpoint = UserInfoAPIEndpoint.fetchProfile(with: requestDTO)
-    return Future<ProfileImageEntity, MainError> { [unowned self] promise in
-      service.request(endpoint: endpoint)
-        .mapError { MainError.networkError($0) }
+    // UserDefaults 확인
+    if let imageURL = MyInfoManager.userProfileURL {
+      return Future { promise in
+        promise(.success(.init(image: imageURL)))
+      }
+    }
+    
+    guard let loggedInUserId = MyInfoManager.id else {
+      return Future { promise in
+        promise(.failure(MainError.referenceError(.other("유저디폴츠에서 로그인한 사용자의 id가져올 수 없습니다."))))
+      }
+    }
+    
+    // 프로필 없는 경우 서버에서 불러오기
+    return Future { [unowned self] promise in
+      othersProfileRepository.fetchProfile(with: loggedInUserId)
         .sink { completion in
           if case .failure(let error) = completion {
             promise(.failure(error))
           }
-        } receiveValue: { responseDTO in
-          let entity = responseDTO.result.toDomain()
-          promise(.success(entity))
+        } receiveValue: { profileImageEntity in
+          MyInfoManager.updateProfileURL(with: profileImageEntity.image)
+          promise(.success(profileImageEntity))
         }.store(in: &subscriptions)
     }
   }
